MIFOSX-2135: Incorporating the feedback

+++ mifosng-provider/src/main/java/org/mifosplatform/infrastructure/configuration/data/ExternalServicesData.java
@@ -2,15 +2,15 @@ package org.mifosplatform.infrastructure.configuration.data;
 
 public class ExternalServicesData {
 	
-	private final int id;
+	private final Long id;
 	private final String name;
 	
-	public ExternalServicesData(final int id, final String name){
+	public ExternalServicesData(final Long id, final String name){
 		this.id =id;
 		this.name = name;		
 	}
 
-	public int getId() {
+	public Long getId() {
 		return id;
 	}
